requirements.txt
openai==1.76.0
python-dotenv==1.1.0
spotipy==2.25.1
